# [FYI, do not merge as-is] Slightly faster variant of Sigma using a dedicated primitive record