--- a/_CoqProject
+++ b/_CoqProject
@@ -2,6 +2,8 @@
 Original.v
 Canonical.v
 Sigma.v
+Record.v
+PrimitiveRecord.v
 GADT.v
 Node01.v
 Patricia.v